updated code with changes in design

+++ src/app/Components/headers/navbar/user-management/user-management.component.ts
@@ -2,14 +2,16 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { CredenceServiceService } from 'src/app/credence-service.service';
 import { Clients, Role, User } from './model/user';
-
+import { ConfirmationService, MessageService } from 'primeng/api';
+import {ConfirmDialogModule} from 'primeng/confirmdialog';
 
 
 
 @Component({
   selector: 'app-user-management',
   templateUrl: './user-management.component.html',
-  styleUrls: ['./user-management.component.css']
+  styleUrls: ['./user-management.component.css'],
+  providers: [ConfirmationService, MessageService]
 })
 export class UserManagementComponent implements OnInit {
 
@@ -24,10 +26,11 @@ export class UserManagementComponent implements OnInit {
   selectedFiles?: FileList;
 
 
-  constructor(private service:CredenceServiceService) { 
+  constructor(private service:CredenceServiceService,private messageService: MessageService,private confirmationService: ConfirmationService) { 
     this.roles=[
       {role:"Admin"},
-      {role:"Super Admin"}
+      {role:"Super Admin"},
+      {role:"PIA"}
     ]
   }
 
@@ -89,29 +92,68 @@ export class UserManagementComponent implements OnInit {
   }
 
   deleteUser(id:string){
-    this.service.deleteUser(id).subscribe(
-      (data:User)=>{
-        console.log("user deleted succsessfully");
-        this.ngOnInit();
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to Delete User?',
+      accept: () => {
+        this.service.deleteUser(id).subscribe(
+          (data:User)=>{
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Deleted',
+              detail: 'user deleted successfully',
+            });
+            this.ngOnInit();
+          },
+          (error:HttpErrorResponse)=>{
+            alert(error);
+          }
+        )
       },
-      (error:HttpErrorResponse)=>{
-        alert(error);
+      reject:()=>{
+        this.messageService.add({
+          severity: 'warn',
+          summary: 'Cancelled',
+          detail: 'user not deleted',
+        });
       }
-    )
+  });
+    
   }
 
   onClickUpdate()
   {
-    this.service.editUserById(this.userData1).subscribe(
-      (data:User)=>{
-        console.log("User updated successfully");
-        this.editUserDialog=false;
-        this.ngOnInit();
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to edit this user?',
+      accept: () => {
+        this.service.editUserById(this.userData1).subscribe(
+          (data:User)=>{
+            console.log("User updated successfully");
+            this.editUserDialog=false;
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Success',
+              detail: 'user updated Successfully',
+            });
+            this.ngOnInit();
+          },
+          (error:HttpErrorResponse)=>{
+            this.messageService.add({
+              severity: 'cancle',
+              summary: 'Cancelled',
+              detail: `${error}`,
+            });
+          }
+        )
       },
-      (error:HttpErrorResponse)=>{
-        alert(error);
+      reject:()=>{
+        this.messageService.add({
+          severity: 'warn',
+          summary: 'Cancelled',
+          detail: 'user not updated',
+        });
       }
-    )
+  });
+    
   }
 
   onClickUpload()
